bert-semantic-search/requirements.txt
transformers
sentence-transformers
faiss-cpu
scikit-learn
matplotlib
pandas
